[CheckVersion] Move update module to generator
